Implement proper parent channel resolution
IMPORTANT: This is disabled for the time being, because it increases the time
it takes to process messages by roughly two orders of magnitude
(4 ms to 600 ms).
It can be enabled in the future, when I have found a better way to do this.

The reason I am committing this at all is because the way it was previously
done was already slow (~300 ms), and this fixes that.

I made a mistake in e823c019f02b9d6a1e04d31b8d71a435e48e2c9c by assuming that
a channel's parent is always a category. That's false, as the parent of a
thread, is a regular channel.
This fixes that.